Redesigned some UI Dependencies
Removed uiUtility class. Moved the generation functions into jcrUI. Moved any get/set functions into a new UIController class, to seperate the concerns.

Need to do the same with IO, but will handle at a later date, when files need to be modified.
